requirements.txt
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
flask==1.1.2
Pillow==7.0.0
torchvision==0.8.1
gunicorn==20.0.4